pyproject.toml
[tool.poetry]
name = "pjecz-portal-notarias"
version = "0.1.0"
description = "Portal de Notarias para entregar edictos y revisar escrituras."
authors = ["Guillermo Valdes <guillermo@movimientolibre.com>"]
license = "AGPL3"
readme = "README.md"
packages = [{include = "portal_notarias"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
cryptography = "^42.0.5"
email-validator = "^2.1.0.post1"
flask = "^3.0.2"
flask-login = "^0.6.3"
flask-moment = "^1.0.5"
flask-sqlalchemy = "^3.1.1"
flask-socketio = "^5.3.6"
flask-wtf = "^1.2.1"
google-auth = "^2.28.1"
google-cloud-secret-manager = "^2.18.2"
google-cloud-storage = "^2.14.0"
google-cloud = "^0.34.0"
gunicorn = "^21.2.0"
hashids = "^1.3.1"
jinja2 = "^3.1.3"
openpyxl = "^3.1.2"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.2"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
rq = "^1.16.0"
sendgrid = "^6.11.0"
sqlalchemy = "^2.0.27"
sqlalchemy-utils = "^0.41.1"
unidecode = "^1.3.8"
uvicorn = "^0.27.1"
werkzeug = "^3.0.1"
wtforms = "^3.1.2"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
pytest = "^8.0.2"
pylint = "^3.1.0"
pylint-sqlalchemy = "^0.3.0"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 128

[tool.isort]
line_length = 128
profile = "black"
